+++ Example/Stopwatch/Sources/ViewControllers/Advanced/Misc/CustomThemesDataSource.h
@@ -1,0 +1,11 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface CustomThemesDataSource : NSObject <UIPickerViewDelegate, UIPickerViewDataSource>
+
+- (NSInteger)currentTheme;
+
+@end
+
+NS_ASSUME_NONNULL_END