+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## 3.26.0
+
+##### Breaking
+- Removed readonly property `overrideApplicationStatusBarHiddenState` in `ABKInAppMessageViewController.h`.
+
+##### Changed
+- Added Binary Project Specification file for more efficient Carthage integration of the full SDK. 
+  - Update your Cartfile to use `binary "https://raw.githubusercontent.com/Appboy/appboy-ios-sdk/master/appboy_ios_sdk_full.json"`
+  - Support for this integration method was added starting with version 3.24.0 of the SDK.
+
+##### Fixed
+- Fixes an issue with in-app messages not respecting the application's status bar style when _View controller-based status bar appearance_ (`UIViewControllerBasedStatusBarAppearance`) is set to `YES` in the Info.plist.
+- Fixes an issue which can lead to text being cut off in Content Cards for specific iPhone models.
+- Fixes an issue preventing test Content Cards from displaying under specific conditions.
+
+#### Added
+- Adds support for specifying `PushStoryAppGroup` in the `Appboy` dictionary in your app's `Info.plist`. This [Apple App Group](https://developer.apple.com/documentation/bundleresources/entitlements/com_apple_security_application-groups?language=objc) will share the [Braze Push Story](https://www.braze.com/docs/developer_guide/platform_integration_guides/ios/push_story/) information such as Campaign IDs between applications from a single Apple Developer account.
+- Adds `appboyBridge.getUser().addAlias(alias, label)` to the javascript interface for HTML in-app messages.
+- Adds the property `overrideUserInterfaceStyle` to `ABKInAppMessage` that allows forcing Light or Dark mode in the same way as Apple's [`UIViewController.overrideUserInterfaceStyle`](https://developer.apple.com/documentation/uikit/uiviewcontroller/3238087-overrideuserinterfacestyle?language=objc).
+  - You can set this property in the `beforeInAppMessageDisplayed:` method of an [ABKInAppMessageControllerDelegate](https://www.braze.com/docs/developer_guide/platform_integration_guides/ios/in-app_messaging/customization/#setting-delegates).
+- Adds the ability to dismiss modal in-app messages when the user clicks outside of the in-app message.
+  - This feature is disabled by default.
+  - You can enable the feature by adding the `Appboy` dictionary to your `Info.plist` file. Inside the `Appboy` dictionary, add the `DismissModalOnOutsideTap` boolean subentry and set the value to `YES`.
+  - You can also enable the feature at runtime by setting `ABKEnableDismissModalOnOutsideTapKey` to `YES` in `appboyOptions`.
+
 ## 3.25.0
 
 ##### Breaking